pyproject.toml
[project]
name = "alx-travel-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-mysql>=4.19.0",
    "djangorestframework>=3.16.1",
    "drf-yasg>=1.21.11",
    "gunicorn>=23.0.0",
    "mysqlclient>=2.2.7",
    "rabbitmq>=0.2.0",
    "sqlparse>=0.5.3",
]

[dependency-groups]
dev = [
    "ruff>=0.14.3",
]
